requirements.txt
django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2